% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_fun.R
\name{axis_fun}
\alias{axis_fun}
\title{Helper function for axis labels using base plotting}
\usage{
axis_fun(
  side,
  at,
  at2,
  labels,
  cex.axis = 1,
  line = -0.25,
  lwd = 1,
  maj.tcl = -0.6,
  min.tcl = -0.3,
  las = 1,
  axisLine = 0
)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be drawn on. The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{at}{major points at which tick-marks are to be drawn. Non-finite (infinite, NaN or NA) values are omitted.}

\item{at2}{minor point at which tick-marks are to be drawn. Non-finite (infinite, NaN or NA) values are omitted.}

\item{labels}{major tick mark labels}

\item{cex.axis}{size of major axis labels text (default = 1)}

\item{line}{the number of lines into the margin at which the axis line will be drawn (default = -0.25)}

\item{lwd}{line widths for the axis line and the tick marks}

\item{maj.tcl}{length of major ticks}

\item{min.tcl}{length of minor ticks}

\item{las}{direction of text}

\item{axisLine}{the number of lines into the margin at which the axis line will be drawn}
}
\value{
a function that provides major and minor ticks with labels
}
\description{
Helper function for axis labels using base plotting
}
\note{
see axis() function
}
\examples{
plot(runif(25),ylim=c(0,1),xlim=c(0,1),xaxt="n",yaxt="n")
axis_fun(1,seq(0,1,0.5),seq(0,1,0.25),seq(0,1,0.5))
axis_fun(2,seq(0,1,0.5),seq(0,1,0.25),seq(0,1,0.5))
}
\keyword{axis}
\keyword{ticks}
