% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweek.period.r
\name{biweek.period}
\alias{biweek.period}
\title{Biweekly Period during Water Year}
\usage{
biweek.period(date, WY.type = "FL")
}
\arguments{
\item{WY.type}{Specifies Federal ("Fed") or Florida ("FL") water years}

\item{Date}{Date in as.POSIXct or as.Date format}
}
\value{
Similar to "Week Number" function in base or lubridate, this function determines the two week period starting during any given water year (Federal or Florida).
}
\description{
Biweekly Period during Water Year
}
\note{
See WY() function
}
\examples{
dates=as.Date(c("2015-05-01","2015-08-20","2015-12-15"))
biweek.period(dates)
}
\keyword{DOWY}
\keyword{biweek}
\keyword{period}
