% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversions}
\alias{conversions}
\alias{cfs.to.m3d}
\title{Random conversion helpers}
\usage{
cfs.to.m3d(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Random conversion helpers
}
\details{
converts data from one unit to another.
\itemize{
\item{cfs.to.m3d}
\item{cfs.to.km3d}
\item{cfs.to.acftd}
\item{mt.to.kg}
\item{kg.to.mt}
\item{acres.to.m2}
\item{m3d.to.cfs}
\item{ft.to.m}
\item{m.to.ft}
\item{degF.to.C}
\item{in.to.cm}
\item{hPa.to.mmHg}
}
}
\examples{
cfs.to.m3d(350)
}
