% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.interp.R
\name{dat.interp}
\alias{dat.interp}
\title{Nutrient Load: Linearly interpolates data to conduct nutrient load calculations}
\usage{
dat.interp(x)
}
\arguments{
\item{x}{numeric}
}
\value{
one of two functions used to calculate daily nutrient load from flow and water quality parameters
}
\description{
Nutrient Load: Linearly interpolates data to conduct nutrient load calculations
}
\examples{
data=c(rep(NA,5),runif(2,2,5),NA,NA,NA,6,7,8, NA,NA)
dat.interp(data)
}
\keyword{Nutrient}
\keyword{load}
