% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_fun.R
\name{date_fun}
\alias{date_fun}
\title{Date function to convert character/other as.POSIXct to date.}
\usage{
date_fun(x, tz = "EST", form = "\%F")
}
\arguments{
\item{x}{as.POSIXct or as.character}

\item{tz}{timezone (default set to Eastern Standard Time)}

\item{form}{format (default set to YYYY-mm-dd)}
}
\value{
Helper function to format date (and date-time) fields
}
\description{
Date function to convert character/other as.POSIXct to date.
}
\examples{
datetime=as.character(c("2015-05-01 08:00:00","2015-10-15 12:00:00","2015-10-15 5:00:00"))
date.fun(datetime);# just date
date.fun(datetime, form="\%F \%X"); #include date and time

date.fun("2020-05-01")
date.fun(as.Date("2020-05-01"))
date.fun(as.POSIXct("2020-05-01 12:00:00",tz="EST"),form="\%F \%H:\%M:\%S")
date.fun(as.POSIXct("2020-05-01 12:00:00",tz="UTC"),form="\%F \%H:\%M:\%S")
}
\keyword{date}
