% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDROmeta.R
\name{DBHYDRO.meta.byDBKEY}
\alias{DBHYDRO.meta.byDBKEY}
\title{South Florida Water Management District online database (DBHYDRO) metadata}
\usage{
DBHYDRO.meta.byDBKEY(DBKEY, ...)
}
\arguments{
\item{DBKEY}{SFWMD station id}

\item{...}{other html parameters}
}
\value{
This function returns water quality dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management. .
}
\description{
South Florida Water Management District online database (DBHYDRO) metadata
}
\examples{
# Water Quality Data
# DBHYDRO.meta.byDBKEY(67486)
}
\keyword{"water}
\keyword{quality"}
