% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDROmeta.R
\name{DBHYDRO.meta.bysite}
\alias{DBHYDRO.meta.bysite}
\title{South Florida Water Management District online database (DBHYDRO) metadata}
\usage{
DBHYDRO.meta.bysite(site, type, cat = "SW", ...)
}
\arguments{
\item{site}{SFWMD station id}

\item{type}{data type ("FLOW","STG","GATE", etc)}

\item{cat}{data category ("SW","RAIN","ETP"); default is "SW"}

\item{...}{other html parameters}
}
\value{
This function returns water quality dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management. .
}
\description{
South Florida Water Management District online database (DBHYDRO) metadata
}
\examples{
# Water Quality Data
# DBHYDRO.meta.bysite("S333","FLOW")
}
\keyword{"water}
\keyword{quality"}
