% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDRO_breakpoint.R
\name{DBHYDRO_breakpoint}
\alias{DBHYDRO_breakpoint}
\title{Data retrievial from South Florida Water Management District online database (DBHYDRO)}
\usage{
DBHYDRO_breakpoint(SDATE, EDATE, DBK, col.names = c("DATETIME",
  "Station", "DBKEY", "Data.Value", "Flag", "Comment"))
}
\arguments{
\item{SDATE}{input start date as.Date (YYYY-MM-DD)}

\item{EDATE}{input end date as.Date (YYYY-MM-DD)}

\item{DBK}{SFWMD daily data site and data type identifier}
}
\value{
This function returns breakpoint (i.e. 15-minute) hydrometerological (discharge, stage(WL) and meterological parameters) dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management.
}
\description{
Data retrievial from South Florida Water Management District online database (DBHYDRO)
}
\examples{
# Daily Discharge Data
sdate=as.Date("2001-05-01");
edate=as.Date("2001-06-01");
dat=DBHYDRO_breakpoint(SDATE,EDATE,"AI516")
}
\keyword{discharge}
\keyword{stage}
\keyword{weather}
