% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDRO_daily.R
\name{DBHYDRO_daily}
\alias{DBHYDRO_daily}
\title{Data retrievial from South Florida Water Management District online database (DBHYDRO)}
\usage{
DBHYDRO_daily(SDATE, EDATE, DBK, offset = 4)
}
\arguments{
\item{SDATE}{input start date as.Date (YYYY-MM-DD)}

\item{EDATE}{input end date as.Date (YYYY-MM-DD)}

\item{DBK}{SFWMD daily data site and data type identifier}

\item{offset}{the values used to remove the data file header.}
}
\value{
This function returns daily hydrometerological (discharge, stage(WL) and meterological parameters) dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management.
}
\description{
Data retrievial from South Florida Water Management District online database (DBHYDRO)
}
\examples{
\dontrun{
# Daily Discharge Data
sdate=as.Date("2001-05-01");
edate=as.Date("2002-05-01");
dat=DBHYDRO_daily(SDATE,EDATE,"FE771")
}
}
\keyword{discharge}
\keyword{stage}
\keyword{weather}
