% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDRO_WQ.R
\name{DBHYDRO_WQ}
\alias{DBHYDRO_WQ}
\title{Data retrieval from South Florida Water Management District online database (DBHYDRO)}
\usage{
DBHYDRO_WQ(
  date_min,
  date_max,
  station_id = NA,
  test_number = NA,
  collect_method = NULL,
  matrix = "SW",
  cust_str = NULL,
  Exclude.FieldQC = T,
  Exclude.Flagged = T,
  sample_type = "SAMP",
  target_code = "file_csv"
)
}
\arguments{
\item{date_min}{input start date as.Date (YYYY-MM-DD)}

\item{date_max}{input end date as.Date}

\item{station_id}{SFWMD Water Quality Monitoring location, see details}

\item{test_number}{corresponds to SFWMD data management system}

\item{collect_method}{default is NULL but can specify G, GP, ACF, ACT, ADT}

\item{matrix}{default is "SW" for surface water}

\item{cust_str}{custom sql query string to add to link (not implemented)}

\item{Exclude.FieldQC}{true/false field, DEFAULT is TRUE}

\item{Exclude.Flagged}{true/false field, DEFAULT is TRUE}

\item{sample_type}{default 'SAMP'}

\item{target_code}{default file_csv but can have "screen" and default browser will launch}
}
\value{
This function returns water quality dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management. .
}
\description{
Data retrieval from South Florida Water Management District online database (DBHYDRO)
}
\details{
Sample Type definitions include SAMP=Sample, RS=Replicate Sample, SS= Split Sample, EB=Equipment Blank, FCEB=Field Cleaned Equipment Blank, FD=Field Duplicate

matrix default is "SW" for surface water other variables including "BAL","BAN","DI","BFE","BFI","GW","PERI","BPL","PW","RA","SE","SO","UNK"

target_code can be either "file_csv" (for CSV) and "screen" (launch default browser). Other formats like "file_txt" (for fixed column width) and "file_pdf" (for pdf) and "screen" (launch default browser) have not been implemented

station_id can use wildcard character but not in combination with wildcard and non-wildcard (i.e. S12\%, S333)
}
\examples{
# Water Quality Data
\dontrun{
sdate=as.Date("2001-05-01");
edate=as.Date("2002-05-01");
parameter=25;#Test Number for Phosphate, Total as P (mg/L)
dat=DBHYDRO_WQ(sdate,edate,"S12\%",parameter,target_code="file_csv")
}
}
\keyword{"water}
\keyword{quality"}
