% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.WY.R
\name{decimal.WY}
\alias{decimal.WY}
\title{Decimal water year from a date field}
\usage{
decimal.WY(date, WY.type = "FL")
}
\arguments{
\item{date}{Date in as.POSIXct or as.Date format}

\item{WY.type}{Specifies Federal ("Fed") or Florida ("FL") water years}
}
\value{
Returns decimal water year value based on Federal (Starts Oct 1) or Florida (Starts May 1),
}
\description{
Decimal water year from a date field
}
\examples{
# Decimal Water Year
decimal.WY(as.Date("2001-05-01"));
}
\keyword{WY}
