% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.Screening.R
\name{DO.Screening}
\alias{DO.Screening}
\title{Function to screen dissolved oxygen data within the Everglades Ecosystem prior to evaluating compliance with established water quality standards.}
\usage{
DO.Screening(data)
}
\arguments{
\item{data}{data.frame() with DateTime, Dissolved Oxygen (concentration) and Temperature data}
}
\value{
DO.Screening() assumes the data has a "DateTime.EST" field (date and time stamp), "DO" Field and "Temp" field to determine suitable data. No data is removed just a "UseData" field is added.
}
\description{
Function to screen dissolved oxygen data within the Everglades Ecosystem prior to evaluating compliance with established water quality standards.
}
\examples{
data=data.frame(DateTime.EST=as.POSIXct(c("2015-05-01 08:00","2015-10-15 12:00","2015-10-15 5:00"),DO=c(35,2,3),Temp=c(25,27,3))
DO.Screening(data)
}
\keyword{Oxygen}
