% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.TOD.WQS.stream.R
\name{DO.TOD.WQS.stream}
\alias{DO.TOD.WQS.stream}
\title{Florida Department of Environmental Protection Time-of-day dissolved oxygen water quality criteria}
\usage{
DO.TOD.WQS.stream(DateTime)
}
\arguments{
\item{DateTime}{Date and Time field in as.POSIXct() format}
}
\value{
DO.TOD.WQS.stream() calculates the time-of-day dissolved oxygen water quality criteria consistent with 62-302.533 Florida Adminstrative Code (FAC) for stream ecosystems. Development of this method can be found at https://floridadep.gov/sites/default/files/tsd-do-criteria-aquatic-life.pdf
}
\description{
Florida Department of Environmental Protection Time-of-day dissolved oxygen water quality criteria
}
\examples{
Date.Time=as.POSIXct(c("2015-05-01 08:00","2015-10-15 12:00"))
DO.TOD.WQS.stream(Date.Time)
}
\keyword{Oxygen}
