% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_PerSat.R
\name{DO_PerSat}
\alias{DO_PerSat}
\title{Calculates dissolved oxygen percent saturation from concentration.}
\usage{
DO_PerSat(Temp, DO, sal)
}
\arguments{
\item{Temp}{temperature (Degree C)}

\item{DO}{Dissolved oxygen (miligrams per liter)}

\item{sal}{Salinity}
}
\value{
DO_PerSat() calculated dissolved oxygen percent saturation consistent with Florida Department of Environmental Protection technical support documents (https://floridadep.gov/sites/default/files/tsd-do-criteria-aquatic-life.pdf) using water temperature, dissolved oxygen concentration and salinity.
}
\description{
Calculates dissolved oxygen percent saturation from concentration.
}
\examples{
DO_PerSat(25.5,15.1,0); #Freshwater
DO_PerSat(25.5,15.1,33); #Marine
}
\keyword{Oxygen}
