% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf_fun.R
\name{ecdf_fun}
\alias{ecdf_fun}
\title{Determines Water Year from a date field}
\usage{
ecdf_fun(x, CI = TRUE, CI.interval = 0.95)
}
\arguments{
\item{x}{numeric vector of the observations for ecdf; for the methods, an object inheriting from class "ecdf".}

\item{CI}{If TRUE CDF confidence intervals will be calculated; default is set to TRUE.}

\item{CI.interval}{Confidence interval; default is set to 95%}
}
\value{
Compute an empirical cumulative distribution function, returns a data.frame() object proportion, value and confidence interval (if CI==T). Code based on base eCDF function.
}
\description{
Determines Water Year from a date field
}
\examples{
set.seed(1)
ecdf_fun(rnorm(100))

}
\keyword{cdf}
