% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars.R
\name{errorbars}
\alias{errorbars}
\title{Errorbars to be used with base plotting}
\usage{
errorbars(x, Mean, SE, col, lty = 1, length = 0.07, lwd = 1)
}
\arguments{
\item{x}{x value}

\item{Mean}{mean value or y alternative}

\item{SE}{standard error}

\item{col}{color of bars}

\item{lty}{Line type The line type. Line types can either be specified as an integer (0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash, 5=longdash, 6=twodash) or as one of the character strings "blank", "solid", "dashed", "dotted", "dotdash", "longdash", or "twodash", where "blank" uses ‘invisible lines’ (i.e., does not draw them).}

\item{length}{length of the edges of the error bar head (in inches).}

\item{lwd}{line width}
}
\value{
draws lines in the form of error bars
}
\description{
Errorbars to be used with base plotting
}
\note{
see arrows() function
}
\examples{
x.val=c(1,2,3)
mean.val=c(2,5,7)
se.val=c(0.1,0.5,1)
plot(mean.val~x.val,ylim=c(0,10))
errorbars(x.val,mean.val,se.val,"red")
}
\keyword{error}
