% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL.Hydroseason.R
\name{FL.Hydroseason}
\alias{FL.Hydroseason}
\title{A function to return the wet/dry hydrologic season specific to south Florida}
\usage{
FL.Hydroseason(Date)
}
\arguments{
\item{Date}{Date in as.POSIXct or as.Date format}
}
\value{
South Florida has a predictable wet and dry season due to the sub-tropical/tropical climate. Florida Department of Environmental Protection South Florida Water Management District and other agencies recognize this seasonality. May to October is considered the wet season and November to Apirl is the dry season for south Florida.
}
\description{
A function to return the wet/dry hydrologic season specific to south Florida
}
\examples{
dates=as.Date(c("2015-05-01","2015-08-20","2015-12-15"))
FL.Hydroseason(dates)
}
\keyword{date}
\keyword{season}
