% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Folder.Maker.R
\name{Folder.Maker}
\alias{Folder.Maker}
\title{Function to make folders within working directory.}
\usage{
Folder.Maker(x)
}
\arguments{
\item{x}{a directory where you want a list of folders created.}
}
\value{
Creates folders and will not overwrite any existing folders
}
\description{
Function to make folders within working directory.
}
\examples{
# paths=paste(getwd(),c("maps","plots","data"),sep="/")
# Folder.Maker(paths)
}
\keyword{files}
