% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.day.R
\name{hydro.day}
\alias{hydro.day}
\title{Day of Water Year}
\usage{
hydro.day(Date, WY.type = "FL")
}
\arguments{
\item{Date}{Date in as.POSIXct or as.Date format}

\item{WY.type}{Specifies Federal ("Fed") or Florida ("FL") water years}
}
\value{
Similar to "Day of the Year" calculations, this functions determines the Day of the Water Year.
}
\description{
Day of Water Year
}
\note{
See WY() function
}
\examples{
dates=as.Date(c("2015-05-01","2015-08-20","2015-12-15"))
hydro.day(dates)
}
\keyword{DOWY}
\keyword{DOY}
