% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHydro_Insight.R
\name{insight_fetch_daily}
\alias{insight_fetch_daily}
\title{Hydrologic and continuous data retrieval from South Florida Water Management District online database (DBHYDRO)}
\usage{
insight_fetch_daily(
  startDate,
  endDate,
  dbkey,
  datum = NULL,
  period = NULL,
  reportType = "timeseries",
  format = "json",
  ...
)
}
\arguments{
\item{startDate}{Required as a Date \code{as.Date} or other format (i.e. \code{as.POSIXct})}

\item{endDate}{Required as a Date \code{as.Date} or other format (i.e. \code{as.POSIXct})}

\item{dbkey}{Required as a character, defines which data is to be accessed. Also called \verb{Timeseries ID}}

\item{datum}{vertical datum of presented data (\code{NGVD29} or \code{NAVD88})}

\item{period}{Timespan of data requested (1day, 3days, 1week, 2 weeks, 30days, 365days, por)}

\item{reportType}{default is \code{timeseries}, Defines the structure of the data presentation see user manual for more}

\item{format}{set to json for ease of data retrieval}

\item{...}{not used, to pass argument along to other functions, in the future.}
}
\description{
using the new data management platform DBHydro Insights,
this function using the Districts API access to retrieve data from the database
}
\examples{
\dontrun{
# Daily Discharge Data
sdate <- as.Date("2001-05-01");
edate <- as.Date("2002-05-01");
dat <- insight_fetch_daily(sdate,edate,"FE771")

# retrieve metadata
attr(dat,"metadata")
}
}
