% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHydro_Insight.R
\name{insight_fetch_ref}
\alias{insight_fetch_ref}
\title{DBHydro Insights reference tables}
\usage{
insight_fetch_ref(table_name)
}
\arguments{
\item{table_name}{reference table name from \url{https://insightsdata.sfwmd.gov/#/reference-tables}}
}
\description{
DBHydro Insights reference tables
}
\examples{
\dontrun{
insight_fetch_ref("collectionMethod")
}
}
