% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_fun.r
\name{leg.fun}
\alias{leg.fun}
\title{Continuous or categorical legend}
\usage{
leg.fun(
  b,
  pal,
  leg.title,
  top.val = 0.8,
  bot.val = 0.2,
  mid.v.val = NULL,
  x.max = 0.3,
  x.min = 0.1,
  mid.val = NULL,
  txt.offset.val = -0.01,
  txt.y = NULL,
  leg.txt = NULL,
  txt.cex = 0.75,
  txt.adj = 0,
  txt.pos = 4,
  txt.offset = 0.5,
  title.cex = 0.8,
  title.pos = 3,
  title.adj = 0,
  title.x = NULL,
  title.y = NULL,
  labs = NULL,
  n.bks = NULL,
  leg.type = c("continuous", "categorical"),
  ...
)
}
\arguments{
\item{b}{breaks or data range}

\item{pal}{color pallette}

\item{leg.title}{legend title}
}
\value{
This function returns water quality dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management. .
}
\description{
Continuous or categorical legend
}
\keyword{"water}
\keyword{quality"}
