% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.Calc.kg.R
\name{Load.Calc.kg}
\alias{Load.Calc.kg}
\title{Nutrient Load: Calculates load from discharge and concentration data}
\usage{
Load.Calc.kg(flow.cfs, conc.mgL)
}
\arguments{
\item{flow.cfs}{Daily discharge data in cubic feet per second}

\item{conc.mgL}{Nutrient (or other parameter) concentration data in milligrams per liter}
}
\value{
Calculates load in kilograms from discharge (cfs) and concentration (mg/L)
}
\description{
Nutrient Load: Calculates load from discharge and concentration data
}
\note{
see dat.interp()
}
\examples{
Load.Calc.kg(350,0.0012);#350 cfs and 0.0012 mg/L
}
\keyword{load}
\keyword{nutrient}
