% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.scale.fun.R
\name{log.scale.fun}
\alias{log.scale.fun}
\title{Generate Log-base 10 sequence of values to be used in conjunction with axis_fun when plotting log data}
\usage{
\method{log}{scale.fun}(val, type)
}
\arguments{
\item{val}{range of values}

\item{type}{major or minor}
}
\value{
a function that provides major and minor ticks with labels on a log scale
}
\description{
Generate Log-base 10 sequence of values to be used in conjunction with axis_fun when plotting log data
}
\note{
see axis() and axis_fun() function
}
\examples{
set.seed(36)
x.val=exp(rlnorm(100))
y.val=exp(rlnorm(100))

xlim.val=c(1,100)
ylim.val=c(1,5000)
plot(y.val~x.val,ylim=ylim.val,xlim=xlim.val,log="xy",yaxt="n",xaxt="n")
axis_fun(1,log.scale.fun(xlim.val,"major"),log.scale.fun(xlim.val,"minor"),log.scale.fun(xlim.val,"major"))
axis_fun(2,log.scale.fun(ylim.val,"major"),log.scale.fun(ylim.val,"minor"),log.scale.fun(ylim.val,"major"))
}
\keyword{axis}
\keyword{log-scale}
\keyword{ticks}
