% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_params.R
\name{model_fit_params}
\alias{model_fit_params}
\title{Model fit evaluation}
\usage{
model_fit_params(Obs, Sim, data, as.DF = T)
}
\arguments{
\item{Obs}{observed data}

\item{Sim}{simulated data}

\item{data}{if data.frame}

\item{as.DF}{default is TRUE and produces data as data.frame(...)}
}
\value{
this function provides various model fit metric including
}
\description{
Model fit evaluation
}
