% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{plot.windrose}
\alias{plot.windrose}
\title{Plot a windrose Object}
\usage{
\method{plot}{windrose}(
  x,
  type = c("count", "mean", "median", "fivenum"),
  convention = c("meteorological", "oceanographic"),
  fill.col = "red",
  bx.border.col = "blue",
  border.col = "darkgrey",
  box.cols = c("red", "pink"),
  lab.col = NULL,
  lab.val = c("S", "W", "N", "E"),
  lab.cex = 1,
  lab.offset = 0.25,
  max.val = NULL,
  med.lwd = 2,
  med.col = "black",
  ...
)
}
\arguments{
\item{x}{a \code{windrose-class} object}

\item{type}{The thing to be plotted, either the number of counts in the angle
interval, the mean of the values in the interval, the median of the values, or
a \code{\link[=fivenum]{fivenum()}} representation of the values.}

\item{convention}{String indicating whether to use meteorological convention or
oceanographic convention for the arrows that emanate from the centre of the
rose.  In meteorological convection, an arrow emanates towards the right on
the diagram if the wind is from the east; in oceanographic convention, such an
arrow indicates flow \emph{to} the east.}

\item{fill.col}{color of fill}

\item{bx.border.col}{box border color}

\item{border.col}{border color}

\item{box.cols}{cox color}

\item{lab.col}{label colors}

\item{lab.val}{label value}

\item{lab.cex}{label size}

\item{lab.offset}{label offset}

\item{max.val}{max value}

\item{med.lwd}{median line width}

\item{med.col}{median color}

\item{...}{to pass arguments to other components of the function}
}
\value{
base plot
}
\description{
plot a windrose objet simialr to \code{oce::windrose.plot} but  a paired down customizable version
}
\examples{
\dontrun{
wind_dat <- windrose(u,v,wnd.dat.xtab)

plot(wind_dat,"fivenum","oceanographic",max.val=round(max(wind_dat$fives,na.rm=T)*1.5),lab.col="black",xpd=NA)
axs.lab=seq(0,round(max(wind_dat$fives,na.rm=T)*1.5),length.out=3)
xx = seq(0,-1,length.out=length(axs.lab))
axis_fun(1,xx,xx,axs.lab,line=-1.2,axisLine=-0.7)
mtext(side=3,adj=0,"LZ40 Wind Data (m s\u207B\u00B9)",line=-1,cex=0.8,font=2)

}
}
