% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_line.R
\name{pt_line}
\alias{pt_line}
\title{Base plotting helper function.}
\usage{
pt_line(
  x,
  y,
  ln.lty,
  ln.col,
  ln.lwd,
  pch,
  bg,
  cex = 1,
  pt.lwd = 0.25,
  pt.lty = 1,
  pt.col = "black"
)
}
\arguments{
\item{x}{x-value}

\item{y}{y-value}

\item{ln.lty}{line type}

\item{ln.col}{line color}

\item{ln.lwd}{line width}

\item{pch}{plotting ‘character’, i.e., symbol to use. see points()}

\item{bg}{background (fill) color for the open plot symbols given by pch = 21:25. see points()}

\item{cex}{point size}

\item{pt.lwd}{line width of points}

\item{pt.lty}{line type for points}

\item{pt.col}{color of point}
}
\value{
draws points and lines together.
}
\description{
Base plotting helper function.
}
\note{
see points() and lines() functions
}
\examples{
x.val=c(1,2,3)
mean.val=c(2,5,7)
se.val=c(0.1,0.5,1)
plot(mean.val~x.val,ylim=c(0,10),type="n")
pt_line(x.val,mean.val,2,"grey",1,21,"dodgerblue1")
}
\keyword{lines}
\keyword{points}
