% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.function.r
\name{qq.function}
\alias{qq.function}
\title{Calculation of theortical quantile (Quantile-Quantile Plot)}
\usage{
qq.function(y)
}
\arguments{
\item{y}{data sample}
}
\value{
Part of the qqnorm generic function
}
\description{
Calculation of theortical quantile (Quantile-Quantile Plot)
}
\examples{
set.seed(123)
x.val=rnorm(100)
y.val=x.val+100
test=lm(y.val~x.val)
qq.function(test$residuals)
}
