% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.access.R
\name{read.access}
\alias{read.access}
\title{Data retrievial from South Florida Water Management District online database (DBHYDRO)}
\usage{
read.access(channel, table)
}
\arguments{
\item{channel}{The path to the access (.accdb) files with extension}

\item{table}{Table within databases. If unsure check RODBC::sqlTables()}
}
\value{
This function allows for quick access of data within an 2007/2010 Microsoft Access database (.accdb).
}
\description{
Data retrievial from South Florida Water Management District online database (DBHYDRO)
}
\examples{

dat=read.access("/database.accdb","WaterQuality")

}
\keyword{access}
