% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalinityCalc.R
\name{SalinityCalc}
\alias{SalinityCalc}
\title{Calculates salinity from Specific Conductance and Temperature data}
\usage{
SalinityCalc(SpCond, Temp, Ref.Cond = 42914)
}
\arguments{
\item{SpCond}{specific conductance (microsiemens per centimeter)}

\item{Temp}{temperature (Degree C)}

\item{Ref.Cond}{The standard reference conductance of seawater (42914 uS/cm)}
}
\value{
Calculation based on UNESCO "Algorithms for computation of fundamental propoerties of seawater".
}
\description{
Calculates salinity from Specific Conductance and Temperature data
}
\examples{
SalinityCalc(50512,25.5)
}
\keyword{Salinity}
