% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE}
\alias{SE}
\title{Calcualtes standard error}
\usage{
SE(x, NA.val = "NA")
}
\arguments{
\item{x}{numeric value}

\item{NA.val}{the value to exclude, default is 'NA'}
}
\value{
Calculates standard error (sd()/sqrt(n)).
}
\description{
Calcualtes standard error
}
\examples{
set.seed(123)
example=c(rnorm(12),NA,NA,NA)
SE(example);#should return 0.2671171

}
\keyword{error}
