% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaded.range.R
\name{shaded.range}
\alias{shaded.range}
\title{Base plotting helper function to draw a shaded polygon}
\usage{
shaded.range(x, y.L, y.U, bg, col = bg, lty = 3, col.adj = 0.25, lwd = 1)
}
\arguments{
\item{x}{x-value}

\item{y.L}{lower y-value}

\item{y.U}{upper y-value}

\item{bg}{background shading}

\item{col}{line color}

\item{lty}{line type}

\item{col.adj}{factor modifying the opacity alpha. See alpha.f in adjustcolor()}

\item{lwd}{lines width around polygon}
}
\value{
draws polygon based on x, y (upper and lower) values
}
\description{
Base plotting helper function to draw a shaded polygon
}
\note{
see polygon() function
}
\examples{
set.seed(456)
y.val=rnorm(100)
x.val=y.val*rnorm(100)
mod=lm(y.val~x.val)
x.pred=seq(min(x.val),max(x.val),max(x.val)/10)
mod.pred=predict(mod, newdata=data.frame(x.val=x.pred),interval="confidence")

plot(y.val~x.val,type="n")
shaded.range(x.pred,mod.pred[,2],mod.pred[,3],"grey")
points(x.val,y.val,pch=19)
}
\keyword{shaded}
