% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{summary.windrose}
\alias{summary.windrose}
\title{Summarize a windrose Object}
\usage{
\method{summary}{windrose}(object, ...)
}
\arguments{
\item{object}{object A \linkS4class{windrose} object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
select summary stats as text
}
\description{
Summarizes some of the data in a \linkS4class{windrose} object, like in oce package
}
