% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thessian_create.R
\name{thessian_create}
\alias{thessian_create}
\title{GIS Function to create thessian polygons from spatial points.}
\usage{
thessian_create(data, clip, plot = T, full.extent = T)
}
\arguments{
\item{data}{spatial dataset}

\item{clip}{polygon (i.e. study area) to clip the thessian polygons.}

\item{plot}{to plot the generate thessian polygons}

\item{full.extent}{extend the thessian polygons to the full extent of the clip-layer.}
}
\value{
Returns a spatial thessian polygon
}
\description{
GIS Function to create thessian polygons from spatial points.
}
