% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{Create a windrose Object}
\usage{
windrose(x, y, data, dtheta = 15)
}
\arguments{
\item{x}{The x component of wind speed, in which case the u and v components of that object are used for the components of wind speed, and y here is ignored.}

\item{y}{The y component of wind speed (or stress).}

\item{data}{Identifies dataframe}

\item{dtheta}{The angle increment (in degrees) within which to classify the data.}
}
\value{
a windrose object, like oce::as.windrose with data slot containing
\itemize{
\item \code{n} - the number of x values
\item \code{x.mean} - the mean of hte x values
\item \code{y.mean} - the mean of the y values
\item \code{theta} - the central angel (in degrees for the class)
\item \code{count} - the number of observation in this class
\item \code{mean} - the mean of the observations in this class
\item \code{fivenum} - the \code{fivenum()} vector for observations in this class (min, the hinge, median, upper hinge and max)
}
}
\description{
Create a wind-rose object
}
\examples{
\dontrun{
library(reshape2)
#' dates <- rep(as.Date(Sys.Date())-1,2)
dbkeys.bk <- data.frame(SITE="LZ40",DBKEY=c("IY031","IY032"),param=c("WNDD","WNDS"))

wnd.dat2 <- insight_fetch_daily(dates[1],dates[2],dbkeys.bk$DBKEY)
wnd.dat2 <- merge(wnd.dat2,dbkeys.bk,"DBKEY")

wnd.dat.xtab <- dcast(wnd.dat2,DATETIME~param,value.var = "Data.Value",mean)
wnd.dat.xtab$WNDS.ms <- wnd.dat.xtab$WNDS*0.44704
wnd.dat.xtab$u <- wnd.dat.xtab$WNDS.ms*cos(wnd.dat.xtab$WNDD*pi/180)
wnd.dat.xtab$v <- wnd.dat.xtab$WNDS.ms*sin(wnd.dat.xtab$WNDD*pi/180)

wind_dat <- windrose(u,v,wnd.dat.xtab)
}
}
