% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WY.R
\name{WY}
\alias{WY}
\title{Determines Water Year from a date field}
\usage{
WY(date, WY.type = "FL")
}
\arguments{
\item{date}{Date in as.POSIXct or as.Date format}

\item{WY.type}{Specifies Federal ("Fed") or Florida ("FL") water years}
}
\value{
Returns water year value based on Federal (Starts Oct 1) or Florida (Starts May 1),
}
\description{
Determines Water Year from a date field
}
\examples{
dates=as.Date(c("2015-05-01","2015-08-20","2015-12-15"))
WY(dates)
}
\keyword{WY}
