% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDROmeta.R
\name{DBHYDRO.meta.bysite}
\alias{DBHYDRO.meta.bysite}
\title{South Florida Water Management District online database (DBHYDRO) metadata}
\usage{
DBHYDRO.meta.bysite(
  site = NA,
  station = NA,
  data_type = NA,
  cat = NA,
  freq = NA,
  stat = NA,
  returnlink = F,
  ...
)
}
\arguments{
\item{site}{SFWMD station id}

\item{cat}{data category ("SW","RAIN","ETP"); default is "SW"}

\item{...}{other html parameters as a list (i.e. list(v_basin="L_OKEE", v_dbkey_list_flag="Y",v_order_by="STATION")}

\item{type}{data type ("FLOW","STG","GATE", etc)}
}
\value{
This function returns water quality dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management. .
}
\description{
South Florida Water Management District online database (DBHYDRO) metadata
}
\examples{
# Water Quality Data
# DBHYDRO.meta.bysite(site="S333",data_type="FLOW")
}
\keyword{"water}
\keyword{quality"}
