% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn.letters.R
\name{dunn.letters}
\alias{dunn.letters}
\title{Plotting helper function to add letters indicating statistically similar parings.}
\usage{
dunn.letters(levels, x, y, labels, col2 = col2, cex2 = cex2)
}
\arguments{
\item{levels}{The number of groups}

\item{x}{a list of x variables}

\item{y}{a list of y variables}

\item{labels}{what text do you want to include}

\item{col2}{color of the text label}

\item{cex2}{size of label}
}
\value{
Helper function to add letters indicating statitcal significance.
}
\description{
Plotting helper function to add letters indicating statistically similar parings.
}
\note{
Use in combination with dunn.test::dunn.test() and rcompanion::cldList()
}
\examples{
## Function id deprecated
}
\keyword{Dunn}
\keyword{boxplot}
\keyword{comparison}
