% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHydro_Insight.R
\name{insight_fetch_wq}
\alias{insight_fetch_wq}
\title{Water Quality Data retrieval from South Florida Water Management District online database (DBHYDRO)}
\usage{
insight_fetch_wq(
  startDate,
  endDate,
  station_id,
  test_number,
  methods = NULL,
  projects = NULL,
  matrices = NULL,
  paramGroups = NULL,
  sampleTypes = NULL,
  reportType = "timeseries",
  format = "csv",
  ...
)
}
\arguments{
\item{startDate}{Required as a Date \code{as.Date} or other format (i.e. \code{as.POSIXct})}

\item{endDate}{Required as a Date \code{as.Date} or other format (i.e. \code{as.POSIXct})}

\item{station_id}{SFWMD Water Quality Monitoring location \url{https://insights.sfwmd.gov/#/mappage/site}}

\item{test_number}{The DBHydro 1 to 4 digit test number, see details}

\item{methods}{A 1 to 3 letter code for the method of sample collection, see details}

\item{projects}{DBHydro water quality project codes}

\item{matrices}{A 1 to 3 letter code for the sampling material}

\item{paramGroups}{Letter code for the parameter group, see details}

\item{sampleTypes}{A letter code defining sample and other type of samples (i.e. QC/QC)}

\item{reportType}{default is \code{timeseries}}

\item{format}{default is \code{csv}}

\item{...}{not used, to pass argument along to other functions, in the future.}
}
\description{
using the new data management platform DBHydro Insights,
this function using the Districts API access to retrieve data from the database
}
\details{
For \code{test_number}, \code{methods},\code{projects},\code{matrices},\code{sampleTypes} and others
see the \code{insight_ref()} function for the corresponding tables.

The function returns a data.frame from DBHydro insights with extract columns, all original data is unaltered.
The following columns have been added to the data.frame
\itemize{
\item \code{colDATETIME} = Collection datetime as a POSIXct object in the America/New_York timezone (EST/EDT)
\item \code{firstTriggerDATETIME} = for autosamplers, First Trigger Date as a POSIXct object in the America/New_York timezone (EST/EDT)
\item \code{censored} = logical \code{TRUE}/\code{FALSE} if the values was reported below the MDL
\item \code{HalfMDL} = if the value was reported less than the MDL than the value was replaced by half the MDL.
}

For \code{paramGroups}
\itemize{
\item O: organics
\item N: nutrients
\item M: metals
\item P: physical parameters
\item B: biological
\item F: field
\item MI: major ions
\item MIS: miscellaneous
}
}
\examples{
\dontrun{
sdate <- as.Date("2001-05-01");
edate <- as.Date("2002-05-01");
dat <- insight_fetch_wq(sdate,edate,c("S12A","S12B"),"25")

# retrieve metadata
attr(dat,"metadata")
}
}
