% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{N}
\alias{N}
\title{Returns sample size}
\usage{
N(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric value}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
return the number of real values.
}
\description{
Returns sample size
}
\examples{
example=c(rnorm(12),NA,NA,NA)
N(example);# versus length(example)
}
\keyword{count}
