% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.function.r
\name{qq.function}
\alias{qq.function}
\title{Calculation of theortical quantile (Quantile-Quantile Plot)}
\usage{
qq.function(y)
}
\arguments{
\item{y}{data sample}
}
\value{
Part of the qqnorm generic function
}
\description{
Calculation of theortical quantile (Quantile-Quantile Plot)
}
\examples{
\donttest{
set.seed(123)
test=lm(rnorm(100)~rnorm(100)+100)
qq.function(test$residuals)
}
}
