% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHydro_Insight.R
\docType{data}
\name{ref_table_name}
\alias{ref_table_name}
\title{A list of reference table names for use with \code{insight_fetch_ref}}
\format{
A character vector of length 28.
}
\usage{
ref_table_name
}
\description{
A character vector of available reference table names as of August 2025 from \url{https://insightsdata.sfwmd.gov/#/reference-tables}.
}
\examples{
\dontrun{
ref_table_name
insight_fetch_ref(ref_table_name[1])  # e.g., "agency"; list of agencies
}
}
\keyword{data}
