% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickplot.dat.arrows.R
\name{stickplot.dat.arrows}
\alias{stickplot.dat.arrows}
\title{windspeed and direction helper functions for base plotting}
\usage{
stickplot.dat.arrows(DateTime, WSPD, WD, data, ...)
}
\arguments{
\item{DateTime}{Date and time field as.POSIXct}

\item{WSPD}{Wind speed in m/s}

\item{WD}{Wind direction in degrees}

\item{data}{data.frame()}
}
\value{
draws line segments in the direct of wind with length corresponding to speed.
}
\description{
windspeed and direction helper functions for base plotting
}
\examples{
#Data from https://www.ndbc.noaa.gov
wx.dat2005=read.table(...)

plot(WSPD~DateTime.EST,wx.dat2005,
type="n",ylab="Wind Speed (m/s)",ylim=c(-20,30))
stickplot.dat.arrows(DateTime.EST,WSPD,WD,wx.dat2005,col="green",lty=1,lwd=1)
}
\keyword{base-plot}
\keyword{wind}
